moving localization story to general, improving addon with newStrings
